AUthDokuWiki: if \OC:$WEBROOT is empty then the cookie-path must be /
Why? Otherwise the cookie-path seems to be constructed as relative to
the current page, which is /apps/dokuwiki/page. However, this cookie is
then not destroyed at logout.

OTOH, we want to keep that cookie as the RPC framework otherwise needs
to relogin on each request.

Alternatives: store the cookies in the session etc.