Rework sizeing of the iframe element.
There was quite a bit of legacy cruft, accumulated over years with trial
and error.

Today we have this Vue-app setup, the main element just dictates the
size. Just use that and be done.

Also: use a ResizeObserver to catch changes. Seems to work like a
charme. We do not care to support stone-age browsers.